backend/requirements.txt
# Google ADK & AI
google-adk>=1.0.0
google-cloud-aiplatform>=1.60.0
google-generativeai>=0.7.0

# Database
google-cloud-firestore>=2.16.0

# MCP Protocol
mcp>=1.0.0

# API Framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9

# Utilities
python-dotenv>=1.0.0
httpx>=0.27.0
asyncio>=3.4.3
structlog>=24.1.0
pytest>=8.2.0
pytest-asyncio>=0.23.0
